+++ src/main/java/com/jaefan/munpyspring/animal/domain/model/AnimalGender.java
@@ -1,0 +1,5 @@
+package com.jaefan.munpyspring.animal.domain.model;
+
+public enum AnimalGender {
+	MALE, FEMALE, UNKNOWN;
+}
