+++ src/main/java/com/jaefan/munpyspring/animal/domain/model/AnimalNeutered.java
@@ -1,0 +1,5 @@
+package com.jaefan.munpyspring.animal.domain.model;
+
+public enum AnimalNeutered {
+	YES, NO, UNKNOWN;
+}
