+++ src/main/java/com/jaefan/munpyspring/animal/domain/model/AnimalType.java
@@ -1,0 +1,5 @@
+package com.jaefan.munpyspring.animal.domain.model;
+
+public enum AnimalType {
+	DOG, CAT, OTHER
+}
