+++ src/main/java/com/jaefan/munpyspring/animal/domain/model/CrawlingCategory.java
@@ -1,0 +1,5 @@
+package com.jaefan.munpyspring.animal.domain.model;
+
+public enum CrawlingCategory {
+	PUBLIC, PROTECTION, RELEASE
+}
