+++ src/main/java/com/jaefan/munpyspring/animal/domain/model/ProtectionStatus.java
@@ -1,0 +1,8 @@
+package com.jaefan.munpyspring.animal.domain.model;
+
+public enum ProtectionStatus {
+	POSTED, // 공고 중
+	PROTECTED, // 보호 중
+	CLOSED, // 공고 종료
+	RELEASED // 보호 종료
+}
