--- src/main/java/com/jaefan/munpyspring/animalRegistration/domain/model/Breed.java
@@ -1,19 +1,0 @@
-package com.jaefan.munpyspring.animalRegistration.domain.model;
-
-import jakarta.persistence.Entity;
-import jakarta.persistence.GeneratedValue;
-import jakarta.persistence.GenerationType;
-import jakarta.persistence.Id;
-import lombok.NoArgsConstructor;
-
-@Entity
-@NoArgsConstructor
-public class Breed {
-	@Id
-	@GeneratedValue(strategy = GenerationType.IDENTITY)
-	private Long id;
-	private String family;
-	private String breedName;
-	private String breedDescription;
-	private String imagePath;
-}
