--- src/main/java/com/jaefan/munpyspring/animalRegistration/domain/model/ProtectionStatus.java
@@ -1,5 +1,0 @@
-package com.jaefan.munpyspring.animalRegistration.domain.model;
-
-public enum ProtectionStatus {
-	ENTER, EXIT
-}
