--- src/main/java/com/jaefan/munpyspring/animalinfo/domain/model/AnimalGender.java
@@ -1,5 +1,0 @@
-package com.jaefan.munpyspring.animalinfo.domain.model;
-
-public enum AnimalGender {
-	MALE, FEMALE, UNKNOWN;
-}
