--- src/main/java/com/jaefan/munpyspring/animalinfo/domain/model/AnimalNeutered.java
@@ -1,5 +1,0 @@
-package com.jaefan.munpyspring.animalinfo.domain.model;
-
-public enum AnimalNeutered {
-	YES, NO, UNKNOWN;
-}
