--- src/main/java/com/jaefan/munpyspring/animalinfo/domain/model/AnimalType.java
@@ -1,5 +1,0 @@
-package com.jaefan.munpyspring.animalinfo.domain.model;
-
-public enum AnimalType {
-	DOG, CAT, OTHER
-}
