--- src/main/java/com/jaefan/munpyspring/animalinfo/domain/model/CrawlingCategory.java
@@ -1,5 +1,0 @@
-package com.jaefan.munpyspring.animalinfo.domain.model;
-
-public enum CrawlingCategory {
-	PUBLIC, PROTECTION, RELEASE
-}
