+++ src/main/java/com/jaefan/munpyspring/common/web/HealthCheckController.java
@@ -1,0 +1,16 @@
+package com.jaefan.munpyspring.common.web;
+
+import org.springframework.http.ResponseEntity;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RestController;
+
+@RestController
+@RequestMapping("/api/v1/health-check")
+public class HealthCheckController {
+
+	@GetMapping
+	public ResponseEntity<Void> HealthCheck() {
+		return ResponseEntity.ok().build();
+	}
+}
