+++ src/main/java/com/jaefan/munpyspring/shelter/domain/model/Shelter.java
@@ -1,33 +1,42 @@
 package com.jaefan.munpyspring.shelter.domain.model;
 
-import jakarta.persistence.Column;
+import com.jaefan.munpyspring.user.domain.model.User;
+
 import jakarta.persistence.Entity;
 import jakarta.persistence.GeneratedValue;
 import jakarta.persistence.GenerationType;
 import jakarta.persistence.Id;
+import jakarta.persistence.JoinColumn;
+import jakarta.persistence.OneToOne;
+import jakarta.validation.constraints.NotBlank;
+import jakarta.validation.constraints.NotNull;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
-import lombok.Getter;
 import lombok.NoArgsConstructor;
-import lombok.ToString;
 
 @Entity
-@Getter
-@ToString
 @Builder
 @NoArgsConstructor
 @AllArgsConstructor
 public class Shelter {
 	@Id
-	@Column(name = "shelter_id")
 	@GeneratedValue(strategy = GenerationType.IDENTITY)
 	private Long id;
 
+	@NotBlank
 	private String name;
 
+	@NotBlank
 	private String address;
 
+	@NotBlank
 	private String owner;
 
-	private String telno;
+	@NotBlank
+	private String telNo;
+
+	@NotNull
+	@OneToOne
+	@JoinColumn(name = "user_id")
+	private User user;
 }
