--- src/main/java/com/jaefan/munpyspring/signUp/domain/model/Shelter.java
@@ -1,36 +1,0 @@
-package com.jaefan.munpyspring.signUp.domain.model;
-
-import jakarta.persistence.Entity;
-import jakarta.persistence.GeneratedValue;
-import jakarta.persistence.GenerationType;
-import jakarta.persistence.Id;
-import jakarta.persistence.JoinColumn;
-import jakarta.persistence.OneToOne;
-import jakarta.validation.constraints.NotBlank;
-import jakarta.validation.constraints.NotNull;
-import lombok.NoArgsConstructor;
-
-@Entity
-@NoArgsConstructor
-public class Shelter {
-	@Id
-	@GeneratedValue(strategy = GenerationType.IDENTITY)
-	private Long id;
-
-	@NotBlank
-	private String name;
-
-	@NotBlank
-	private String address;
-
-	@NotBlank
-	private String owner;
-
-	@NotBlank
-	private String telNo;
-
-	@NotNull
-	@OneToOne
-	@JoinColumn(name = "user_id")
-	private User user;
-}
