--- src/test/java/com/jaefan/munpyspring/MunpySpringApplicationTests.java
@@ -1,13 +1,0 @@
-package com.jaefan.munpyspring;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class MunpySpringApplicationTests {
-
-	@Test
-	void contextLoads() {
-	}
-
-}
